fix some typos

+++ libraries/Wire/src/Wire.cpp
@@ -132,17 +132,17 @@ uint8_t TwoWire::requestFrom(uint32_t address, uint8_t quantity, uint32_t iaddre
 
 uint8_t TwoWire::requestFrom(uint32_t address, uint8_t quantity, uint8_t sendStop)
 {
-  return requestFrom((uint8_t)address, (uint8_t)quantity, (uint32_t)0, (uint8_t)0, (uint8_t)sendStop);
+  return requestFrom((uint32_t)address, (uint8_t)quantity, (uint32_t)0, (uint8_t)0, (uint8_t)sendStop);
 }
 
 uint8_t TwoWire::requestFrom(uint32_t address, uint8_t quantity)
 {
-  return requestFrom((uint8_t)address, (uint8_t)quantity, (uint8_t)true);
+  return requestFrom((uint32_t)address, (uint8_t)quantity, (uint8_t)true);
 }
 
 uint8_t TwoWire::requestFrom(int address, int quantity)
 {
-  return requestFrom((uint8_t)address, (uint8_t)quantity, (uint8_t)true);
+  return requestFrom((uint32_t)address, (uint8_t)quantity, (uint8_t)true);
 }
 
 uint8_t TwoWire::requestFrom(int address, int quantity, int sendStop)
